+++ src/test/java/com/klarna/hiverunner/SchemaResetBetweenTestMethodsTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2013-2018 Klarna AB
+ * Copyright (C) 2013-2021 Klarna AB
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,9 +16,9 @@
 package com.klarna.hiverunner;
 
 import com.klarna.hiverunner.annotations.HiveSQL;
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -31,7 +31,7 @@ import java.util.List;
  * <p/>
  * This is solved by doing a 'use default' in {@link com.klarna.hiverunner.builder.HiveShellTearable#tearDown()}
  */
-@RunWith(StandaloneHiveRunner.class)
+@ExtendWith(HiveRunnerExtension.class)
 public class SchemaResetBetweenTestMethodsTest {
 
 
@@ -56,7 +56,7 @@ public class SchemaResetBetweenTestMethodsTest {
         Collections.sort(expectedDatabases);
         Collections.sort(actualDatabases);
 
-        Assert.assertEquals(expectedDatabases, actualDatabases);
+        Assertions.assertEquals(expectedDatabases, actualDatabases);
 
     }
 
@@ -74,7 +74,7 @@ public class SchemaResetBetweenTestMethodsTest {
         Collections.sort(expectedDatabases);
         Collections.sort(actualDatabases);
 
-        Assert.assertEquals(expectedDatabases, actualDatabases);
+        Assertions.assertEquals(expectedDatabases, actualDatabases);
     }