+++ src/test/java/com/klarna/hiverunner/io/IgnoreClosePrintStreamTest.java
@@ -20,12 +20,12 @@ import static org.mockito.Mockito.verify;
 
 import java.io.PrintStream;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
+import org.mockito.junit.jupiter.MockitoExtension;
 
-@RunWith(MockitoJUnitRunner.class)
+@ExtendWith(MockitoExtension.class)
 public class IgnoreClosePrintStreamTest {
 
     @Mock